pyproject.toml
[project]
name = "api-mocker"
version = "0.1.1"
description = "The industry-standard, production-ready, free API mocking and development acceleration tool."
authors = [
    { name = "sherin joseph roy", email = "sherin.joseph2217@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["api", "mock", "mocking", "development", "testing", "openapi", "swagger", "cli", "devops"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "typer>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "PyYAML>=6.0",
    "toml>=0.10.2",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/Sherin-SEF-AI/api-mocker"
Repository = "https://github.com/Sherin-SEF-AI/api-mocker.git"

[project.scripts]
api-mocker = "api_mocker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta" 